# swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Student-Supervisor Relationships",
        "title": "SSR Application",
        "contact": {
            "name": "Ivakhnenko Dmitry, Moscow",
            "url": "github.com/ivaahn/ssr",
            "email": "ivahnencko01@gmail.com"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/supervisor/bid": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supervisor"
                ],
                "summary": "GetUserInfo supervisor's bids",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Supervisor ID",
                        "name": "supervisor_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SvBids"
                        }
                    }
                }
            }
        },
        "/api/supervisor/bid/resolve": {
            "post": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supervisor"
                ],
                "summary": "Accept or Decline student's bid",
                "parameters": [
                    {
                        "description": "bid info",
                        "name": "ResolveBid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResolveBid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResolveBidResp"
                        }
                    }
                }
            }
        },
        "/api/supervisor/profile": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supervisor"
                ],
                "summary": "GetUserInfo supervisor's profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SvProfile"
                        }
                    }
                }
            }
        },
        "/api/supervisor/work": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supervisor"
                ],
                "summary": "GetUserInfo supervisor's works",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Supervisor ID",
                        "name": "supervisor_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SvWorkPlenty"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ResolveBid": {
            "type": "object",
            "properties": {
                "accept": {
                    "type": "boolean"
                },
                "bidID": {
                    "type": "integer"
                },
                "supervisorID": {
                    "type": "integer"
                }
            }
        },
        "dto.ResolveBidResp": {
            "type": "object",
            "properties": {
                "new_status": {
                    "type": "string"
                }
            }
        },
...
        "dto.StProfile": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "studentCard": {
                    "type": "string"
                },
                "studentID": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.SubjectResp": {
            "type": "object",
            "properties": {
                "department": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subjectID": {
                    "type": "integer"
                }
            }
        },
        "dto.SvBid": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "student": {
                    "$ref": "#/definitions/dto.StProfile"
                },
                "work": {
                    "$ref": "#/definitions/dto.Work"
                }
            }
        },
        "dto.SvBids": {
            "type": "object",
            "properties": {
                "bids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SvBid"
                    }
                }
            }
        },
        "dto.SvProfile": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "avatarUrl": {
                    "type": "string"
                },
                "birthdate": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "supervisorID": {
                    "type": "integer"
                }
            }
        },
        "dto.SvWork": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "head": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
        "dto.SvWorkPlenty": {
            "type": "object",
            "properties": {
                "supervisorID": {
                    "type": "integer"
                },
                "works": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SvWork"
                    }
                }
            }
        },
        "dto.Work": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "semester": {
                    "type": "integer"
                },
                "subject": {
                    "$ref": "#/definitions/dto.SubjectResp"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Auth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}